wuappa-backend/requirements.txt
Django==1.11.7
django-allauth==0.34.0
django-cities-light==3.3.0
django-filter==1.1.0
django-language-field==0.0.3
django-model-utils==3.0.0
django-rest-auth==0.9.2
djangorestframework==3.7.1
djangorestframework-link-header-pagination==0.1.1
drfdocs==0.0.11
psycopg2==2.7.3.2
djangorestframework-jwt==1.11
django-cors-headers==2.1.0
gunicorn==19.7.1
django-iban==0.3.1
phonenumbers==8.8.5
django-language-field==0.0.3
requests==2.18.4
requests-mock==1.3.0
django-money==0.12.1
django-admin-rangefilter==0.3.0
dj-stripe==1.0.0.post1
django-push-notifications==1.5.0
python-dateutil==2.6.1
django-picklefield==1.0.0
django-constance==2.0.0
django-adminplus==0.5
django-modeltranslation==0.12.2
jsonfield==2.0.2
simplejson==3.13.2
stripe==1.82.2
python-dotenv
